+++ src/AntDesign.Charts/Components/Plots/Liquid/LiquidConfig.cs
@@ -13,11 +13,11 @@ namespace AntDesign.Charts
         [JsonPropertyName("liquidSize")]
         public int? LiquidSize { get; set; }
         [JsonPropertyName("min")]
-        public int? Min { get; set; }
+        public decimal? Min { get; set; }
         [JsonPropertyName("max")]
-        public int? Max { get; set; }
+        public decimal? Max { get; set; }
         [JsonPropertyName("value")]
-        public int? Value { get; set; }
+        public decimal? Value { get; set; }
         [JsonPropertyName("liquidStyle")]
         public object LiquidStyle { get; set; }
         [JsonPropertyName("renderer")]
@@ -95,11 +95,11 @@ namespace AntDesign.Charts
         [JsonPropertyName("liquidSize")]
         public int? LiquidSize { get; set; }
         [JsonPropertyName("min")]
-        public int? Min { get; set; }
+        public decimal? Min { get; set; }
         [JsonPropertyName("max")]
-        public int? Max { get; set; }
+        public decimal? Max { get; set; }
         [JsonPropertyName("value")]
-        public int? Value { get; set; }
+        public decimal? Value { get; set; }
         [JsonPropertyName("liquidStyle")]
         public object LiquidStyle { get; set; }//OneOf <LiquidStyle, ((...args: any[]) => LiquidStyle)>
